Refactor to move Color widgets in separate files
