+++ wrapper/Widgets.cpp
@@ -173,26 +173,6 @@ IggBool iggInputInt(char const *label, int *value, int step, int step_fast, int 
    return ImGui::InputInt(label, value, step, step_fast, flags) ? 1 : 0;
 }
 
-IggBool iggColorEdit3(char const *label, float *col, int flags)
-{
-   return ImGui::ColorEdit3(label, col, flags) ? 1 : 0;
-}
-
-IggBool iggColorEdit4(char const *label, float *col, int flags)
-{
-   return ImGui::ColorEdit4(label, col, flags) ? 1 : 0;
-}
-
-IggBool iggColorPicker3(char const *label, float *col, int flags)
-{
-   return ImGui::ColorPicker3(label, col, flags) ? 1 : 0;
-}
-
-IggBool iggColorPicker4(char const *label, float *col, int flags)
-{
-   return ImGui::ColorPicker4(label, col, flags) ? 1 : 0;
-}
-
 IggBool iggCollapsingHeader(const char *label, int flags)
 {
    return ImGui::CollapsingHeader(label, flags) ? 1 : 0;
